Restructure attestation client and rename string_obligation to string
Reorganize attestation client to match Rust SDK nested structure with
escrow/v1/non_tierable and escrow/v2/non_tierable submodules. Add util
module for get_attestation, register_schema, and attest operations.
Rename string_obligation.rs to string.rs for consistency with Rust SDK.
