+++ sdks/py/src/clients/obligations/mod.rs
@@ -6,13 +6,13 @@
 //! - `erc1155`: ERC1155 token obligations (escrow, payment, barter)
 //! - `token_bundle`: Token bundle obligations (escrow, payment, barter)
 //! - `attestation`: Attestation obligations
-//! - `string_obligation`: String-based obligations
+//! - `string`: String-based obligations
 
 pub mod attestation;
 pub mod erc1155;
 pub mod erc20;
 pub mod erc721;
-pub mod string_obligation;
+pub mod string;
 pub mod token_bundle;
 
 // Re-export main client types
@@ -20,11 +20,11 @@ pub use attestation::AttestationClient;
 pub use erc1155::Erc1155Client;
 pub use erc20::Erc20Client;
 pub use erc721::Erc721Client;
-pub use string_obligation::StringObligationClient;
+pub use string::StringObligationClient;
 pub use token_bundle::TokenBundleClient;
 
 // Re-export obligation data types
 pub use erc1155::{PyERC1155EscrowObligationData, PyERC1155PaymentObligationData};
 pub use erc20::{PyERC20EscrowObligationData, PyERC20PaymentObligationData};
 pub use erc721::{PyERC721EscrowObligationData, PyERC721PaymentObligationData};
-pub use string_obligation::PyStringObligationData;
+pub use string::PyStringObligationData;
