+++ sdks/py/src/clients/obligations/string.rs
